# Optimize clsx implementation for better performance